# Speedup the 2nd Angular server connection onwards