* Set lsp-clients-angular-language-server-command after the first run
to avoid the lag introduced by running the shell command.

Previously, if I add `benchmark-progn` after `lsp-stdio-connection                   (lambda ()`

Then after each `revert-buffer` & new file opened:
```
Elapsed time: 0.221614s
Elapsed time: 0.220887s
Elapsed time: 0.228161s
```

Certainly it's not a good experience for users.

Now:

First connection:
```
Elapsed time: 0.220515s
Elapsed time: 0.000001s [2 times]
```
After that, for each `revert-buffer` & new file opened:
```
Elapsed time: 0.000002s [2 times]
Elapsed time: 0.000001s
```
That's a huge speedup!

* Make Angular server's NodeJS prefix path configurable.

Because sometimes we don't use npm, and maybe there are multiple
installed versions of NodeJS, too.